requirements.txt
django
django-tinymce
cloudinary
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter
djangorestframework-simplejwt
drf-yasg
setuptools
asgiref
autopep8
bootstrap4
certifi
cffi
chardet
cloudinary
confusable-homoglyphs
coreapi
coreschema
cryptography
dj-database-url
django
django-bootstrap3
django-braces
django-cors-headers
django-countries
django-crispy-forms
django-filter
django-generic-aggregation
django-heroku
django-model-utils
django-readonly-field
django-registration
django-rest-knox
django-rql
django-simple-ratings
django-star-ratings
django-tinymce
djangorestframework
djangorestframework-filters
djangorestframework-jwt
djangorestframework-simplejwt
drf-yasg
gunicorn
idna
importlib-metadata
inflection
itypes
Jinja2
lark-parser
MarkupSafe
packaging
Pillow
psycopg2
pycodestyle
pycparser
PyJWT
pyparsing
python-decouple
pytz
requests
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
swapper
toml
uritemplate
urllib3
whitenoise
zipp